requirements.txt
python-dotenv
langchain
langchain-openai
langchain-community
streamlit
pypdf
faiss-cpu
langdetect